// SofTec_TMS/NWC.Web/src/app/TMS-Module/Models/print-customer-invoice.ts
export class PrintCustomerInvoiceDTO {

    CustomerName: string;
    CustomerMobile: string;
    CustomerAddress: string;
    OrderNumber: string
    OrderQuantity: number;
    TotalCost: number;
    ScheduledDeliveryTime: Date;
    EncryptedConfirmationCode: string;
    ZoneName: string;
    ContractorFullName: string;
    TaxNumber: string;
    CompanyAddress: string;
    assignedStation: string;
    plateNo: string;
    StationName: string;
    VehicleID: string;
    NetCost: number;
    InvoiceNo: string;
  TransporterCode: string;
  CategoryName: string;
  Category: string;
  CategoryID: string;
}